move poll out of control
